requirements.txt
alembic==1.4.3
Authlib==0.15.3
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constants==0.6.0
cryptography==3.4.6
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.3.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-jose-cryptodome==1.3.2
pytz==2020.5
requests==2.25.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
